# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  tag.CreateTagInput:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  vo.BasicResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 一个基于 Gin 的论坛系统 API
  title: Tiny Forum API
  version: "1.0"
paths:
...
  /statistics/day:
    get:
      consumes:
      - application/json
      description: 获取指定日期的统计数据（用户、帖子、评论等）
      parameters:
      - default: 今天
        description: '日期 (格式: 2006-01-02)'
        in: query
        name: date
        type: string
      - default: all
        description: 统计类型
        enum:
        - users
        - posts
        - comments
        - all
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取日统计数据
      tags:
      - 统计管理
...
  /statistics/total:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的汇总统计数据
      parameters:
      - description: '开始日期 (格式: 2006-01-02，默认30天前)'
        in: query
        name: start_date
        type: string
      - description: '结束日期 (格式: 2006-01-02，默认今天)'
        in: query
        name: end_date
        type: string
      - default: all
        description: 统计类型
        enum:
        - users
        - posts
        - comments
        - all
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取总计统计数据
      tags:
      - 统计管理
  /statistics/trend:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的趋势数据（按天、周、月统计）
      parameters:
      - description: '开始日期 (格式: 2006-01-02，默认7天前)'
        in: query
        name: start_date
        type: string
      - description: '结束日期 (格式: 2006-01-02，默认今天)'
        in: query
        name: end_date
        type: string
      - description: 统计类型
        enum:
        - users
        - posts
        - comments
        in: query
        name: type
        required: true
        type: string
      - default: day
        description: 统计粒度
        enum:
        - day
        - week
        - month
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取趋势统计数据
      tags:
      - 统计管理
  /tags:
    get:
      description: 获取系统中所有的标签列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      summary: 获取所有标签
      tags:
      - 标签管理
    post:
      consumes:
      - application/json
      description: 创建一个新的标签，需要管理员权限
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建标签（仅管理员）
      tags:
      - 标签管理
  /tags/{id}:
    delete:
      description: 删除指定标签，需要管理员权限
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 无效的标签ID
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除标签（仅管理员）
      tags:
      - 标签管理
    get:
      description: 根据标签ID获取标签信息
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      summary: 获取单个标签
      tags:
      - 标签管理
    put:
      consumes:
      - application/json
      description: 更新指定标签的信息，需要管理员权限
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 请求参数错误或无效的标签ID
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新标签（仅管理员）
      tags:
      - 标签管理
  /timeline/following:
    get:
      description: 获取当前用户关注的人的内容时间线
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取关注时间线
      tags:
      - 时间线管理
  /timeline/home:
    get:
      description: 获取当前用户首页的时间线（推荐内容）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取首页时间线
      tags:
      - 时间线管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"